# frontend/src/generated/openapi.json
{"openapi": "3.0.2", "info": {"title": "Holocron API", "description": "A collection of tools and resources to aid the GM in a campaign of SWRPG", "version": "0.0.1"}, "paths": {"/data/talents": {"get": {"tags": ["data"], "summary": "List Talents", "operationId": "list_talents", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response List Talents Data Talents Get", "type": "array", "items": {"$ref": "#/components/schemas/TalentSchema"}}}}}}}}, "/data/abilities": {"get": {"tags": ["data"], "summary": "List Abilities", "operationId": "list_abilities", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "deprecated": true}}, "/data/skills": {"get": {"tags": ["data"], "summary": "List Skills", "operationId": "list_skills", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response List Skills Data Skills Get", "type": "array", "items": {"$ref": "#/components/schemas/SkillSchema"}}}}}}}}, "/data/force": {"get": {"tags": ["data"], "summary": "List Force Powers", "operationId": "list_force_powers", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "deprecated": true}}, "/data/player/careers": {"get": {"tags": ["data"], "summary": "List Careers", "operationId": "list_careers", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "deprecated": true}}, "/data/player/careers/specializations": {"get": {"tags": ["data"], "summary": "List Career Specializations", "operationId": "list_career_specializations", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "deprecated": true}}, "/data/player/hooks/": {"get": {"tags": ["data"], "summary": "List Hooks", "description": "Hooks, obligations, duties, moralities and motivations", "operationId": "list_hooks", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "deprecated": true}}, "/data/rules/": {"get": {"tags": ["data"], "summary": "List Special Rules", "operationId": "list_special_rules", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "deprecated": true}}, "/data/equipment/weapon": {"get": {"tags": ["data"], "summary": "List Weapons", "operationId": "list_weapons", "parameters": [{"required": false, "schema": {"title": "Categories", "type": "string"}, "name": "categories", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response List Weapons Data Equipment Weapon Get", "type": "array", "items": {"$ref": "#/components/schemas/WeaponSchema"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/data/equipment/weapon/categories": {"get": {"tags": ["data"], "summary": "List Weapon Categories", "operationId": "list_weapon_categories", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/data/equipment/armor": {"get": {"tags": ["data"], "summary": "List Armor", "operationId": "list_armor", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response List Armor Data Equipment Armor Get", "type": "array", "items": {"$ref": "#/components/schemas/ArmorSchema"}}}}}}}}, "/data/equipment/gear": {"get": {"tags": ["data"], "summary": "List Gear", "operationId": "list_gear", "parameters": [{"required": false, "schema": {"title": "Categories", "type": "string"}, "name": "categories", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response List Gear Data Equipment Gear Get", "type": "array", "items": {"$ref": "#/components/schemas/GearSchema"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/data/equipment/gear/categories": {"get": {"tags": ["data"], "summary": "List Gear Categories", "operationId": "list_gear_categories", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response List Gear Categories Data Equipment Gear Categories Get", "type": "array", "items": {"type": "string"}}}}}}}}, "/data/equipment/attachments": {"get": {"tags": ["data"], "summary": "List Attachments", "operationId": "list_attachments", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response List Attachments Data Equipment Attachments Get", "type": "array", "items": {"$ref": "#/components/schemas/AttachmentSchema"}}}}}}}}, "/data/crafting/templates": {"get": {"tags": ["data"], "summary": "List Crafting Templates", "operationId": "list_crafting_templates", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "deprecated": true}}, "/data/dice/tables/": {"get": {"tags": ["data"], "summary": "List Dice Tables", "description": "List tables for spending x, y, a, t on [something]", "operationId": "list_dice_tables", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "deprecated": true}}, "/data/adversaries": {"get": {"tags": ["data"], "summary": "List Adversaries", "operationId": "list_adversaries", "parameters": [{"required": false, "schema": {"title": "Minionsonly", "type": "boolean", "default": false}, "name": "minionsOnly", "in": "query"}, {"required": false, "schema": {"title": "Rivalsonly", "type": "boolean", "default": false}, "name": "rivalsOnly", "in": "query"}, {"required": false, "schema": {"title": "Nemesisonly", "type": "boolean", "default": false}, "name": "nemesisOnly", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "deprecated": true}}, "/data/vehicles/": {"get": {"tags": ["data"], "summary": "List Vehicles", "operationId": "list_vehicles", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "deprecated": true}}, "/data/vehicles/modifications": {"get": {"tags": ["data"], "summary": "List Vehicle Modifications", "operationId": "list_vehicle_modifications", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response List Vehicle Modifications Data Vehicles Modifications Get", "type": "array", "items": {"$ref": "#/components/schemas/AttachmentSchema"}}}}}}}}, "/data/vehicles/weapons": {"get": {"tags": ["data"], "summary": "List Vehicle Weapons", "operationId": "list_vehicle_weapons", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response List Vehicle Weapons Data Vehicles Weapons Get", "type": "array", "items": {"$ref": "#/components/schemas/WeaponSchema"}}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"ArmorSchema": {"title": "ArmorSchema", "required": ["name", "description", "price", "restricted", "rarity", "defense", "soak", "base_mods", "models", "source"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "description": {"title": "Description", "type": "string"}, "price": {"title": "Price", "type": "integer"}, "restricted": {"title": "Restricted", "type": "boolean"}, "rarity": {"title": "Rarity", "type": "integer"}, "defense": {"title": "Defense", "type": "integer"}, "soak": {"title": "Soak", "type": "integer"}, "base_mods": {"title": "Base Mods", "type": "array", "items": {"type": "string"}}, "models": {"title": "Models", "type": "array", "items": {"type": "string"}}, "source": {"title": "Source", "type": "array", "items": {"type": "string"}}}}, "AttachmentSchema": {"title": "AttachmentSchema", "required": ["type", "name", "description", "price", "hp", "restricted", "rarity", "models", "base_mods", "source"], "type": "object", "properties": {"type": {"title": "Type", "type": "string"}, "name": {"title": "Name", "type": "string"}, "description": {"title": "Description", "type": "string"}, "price": {"title": "Price", "type": "integer"}, "hp": {"title": "Hp", "type": "integer"}, "restricted": {"title": "Restricted", "type": "boolean"}, "rarity": {"title": "Rarity", "type": "integer"}, "models": {"title": "Models", "type": "array", "items": {"type": "string"}}, "base_mods": {"title": "Base Mods", "type": "array", "items": {"type": "string"}}, "source": {"title": "Source", "type": "array", "items": {"type": "string"}}}}, "GearSchema": {"title": "GearSchema", "required": ["name", "description", "category", "price", "restricted", "rarity", "models", "source"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "description": {"title": "Description", "type": "string"}, "category": {"title": "Category", "type": "string"}, "price": {"title": "Price", "type": "integer"}, "restricted": {"title": "Restricted", "type": "boolean"}, "rarity": {"title": "Rarity", "type": "integer"}, "models": {"title": "Models", "type": "array", "items": {"type": "string"}}, "source": {"title": "Source", "type": "array", "items": {"type": "string"}}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "SkillSchema": {"title": "SkillSchema", "required": ["name", "characteristic", "type", "description"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "characteristic": {"title": "Characteristic", "type": "string"}, "type": {"title": "Type", "type": "string"}, "description": {"title": "Description", "type": "string"}}}, "Source": {"title": "Source", "required": ["text"], "type": "object", "properties": {"text": {"title": "Text", "type": "string"}, "page": {"title": "Page", "type": "integer"}}}, "TalentSchema": {"title": "TalentSchema", "required": ["name", "description", "ranked", "activation"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "description": {"title": "Description", "type": "string"}, "ranked": {"title": "Ranked", "type": "boolean"}, "activation": {"title": "Activation", "type": "string"}, "sources": {"title": "Sources", "type": "array", "items": {"$ref": "#/components/schemas/Source"}}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}, "WeaponSchema": {"title": "WeaponSchema", "required": ["name", "description", "models", "type", "skill", "price", "restricted", "hp", "rarity", "encumbrance", "damage", "plus_damage", "crit", "range", "qualities", "base_mods", "source"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "description": {"title": "Description", "type": "string"}, "models": {"title": "Models", "type": "array", "items": {"type": "string"}}, "type": {"title": "Type", "type": "string"}, "skill": {"title": "Skill", "type": "string"}, "price": {"title": "Price", "type": "integer"}, "restricted": {"title": "Restricted", "type": "boolean"}, "hp": {"title": "Hp", "type": "integer"}, "rarity": {"title": "Rarity", "type": "integer"}, "encumbrance": {"title": "Encumbrance", "type": "integer"}, "damage": {"title": "Damage", "type": "integer"}, "plus_damage": {"title": "Plus Damage", "type": "boolean"}, "crit": {"title": "Crit", "type": "integer"}, "range": {"title": "Range", "type": "string"}, "qualities": {"title": "Qualities", "type": "array", "items": {"type": "string"}}, "base_mods": {"title": "Base Mods", "type": "array", "items": {"type": "string"}}, "source": {"title": "Source", "type": "array", "items": {"type": "string"}}}}}}, "servers": [{"url": "/"}]}